# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asset Management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "List employees",
                "operationId": "b9b73e1873c5fe36b951f014f711e4fa",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create employee",
                "operationId": "781d9d682a574c5e2c1dab17f782b5cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "name",
                                    "department_id",
                                    "tgl_masuk",
                                    "email"
                                ],
                                "properties": {
                                    "nik": {
                                        "type": "string",
                                        "example": "EMP001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tgl_masuk": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-01-15"
                                    },
                                    "tgl_resign": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "tempat_lahir": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "tgl_lahir": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-04-20"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "budi@example.com"
                                    },
                                    "no_tlp": {
                                        "type": "string",
                                        "example": "+628123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee detail",
                "operationId": "de09c8d9afb3f7114d227ad0645940ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update employee",
                "operationId": "55e1ceec679a0d1436c869735e984a83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Employee"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete employee",
                "operationId": "dc1326c2f6b4bed0201d62396ddfd970",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee deleted"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Employee": {
                "title": "Employee",
                "required": [
                    "nik",
                    "name",
                    "department_id",
                    "tgl_masuk",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nik": {
                        "type": "string",
                        "example": "EMP001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "tgl_masuk": {
                        "type": "string",
                        "format": "date",
                        "example": "2020-01-15"
                    },
                    "tgl_resign": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-05-01",
                        "nullable": true
                    },
                    "tempat_lahir": {
                        "type": "string",
                        "example": "Jakarta"
                    },
                    "tgl_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "1995-04-20"
                    },
                    "email": {
                        "type": "string",
                        "example": "budi@example.com"
                    },
                    "no_tlp": {
                        "type": "string",
                        "example": "+628123456789"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}